Completed Exercises 4.15-4.23

+++ part4/blog/controllers/blogs.js
@@ -3,7 +3,7 @@ const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({});
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 });
     response.json(blogs);
 });
 
@@ -16,18 +16,31 @@ blogsRouter.post('/', async (request, response) => {
         });
     }
 
+    const user = request.user;
+
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes || 0
+        likes: body.likes || 0,
+        user: user.id
     });
 
     const result = await blog.save();
+    user.blogs = user.blogs.concat(result._id);
+    await user.save();
+
     response.status(201).json(result);
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const user = request.user;
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog.user.toString() !== user.id.toString()) {
+        return response.status(401).json({ error: 'this blog can only be deleted by its author' });
+    }
+
     const result = await Blog.findByIdAndRemove(request.params.id);
     if (result) {
         response.status(204).end(); // successfully found blog
@@ -43,7 +56,8 @@ blogsRouter.put('/:id', async (request, response) => {
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes
+        likes: body.likes,
+        user: body.user
     };
 
     const updatedBlog = await Blog.findByIdAndUpdate(
